+++ src/global.h
@@ -2,7 +2,7 @@
 #define __global_h__
 extern bool is_abort;
 extern bool is_parachute_deployed;
-extern int g_servo_pitch;
-extern int g_servo_roll;
+extern int16_t g_servo_pitch;
+extern int16_t g_servo_roll;
 
 #endif
